Change the way file loading works for the taxonomy configuration files, to match the existing code. In doing so, move the Tomcat file classes to the rifGenericLibrary module, which seems slightly more appropriate anyway.
