+++ rifGenericLibrary/src/main/java/org/sahsu/rif/generic/fileformats/tomcat/TomcatResourceBundle.java
@@ -1,4 +1,4 @@
-package org.sahsu.rif.services.system.files;
+package org.sahsu.rif.generic.fileformats.tomcat;
 
 import org.sahsu.rif.generic.util.RIFLogger;
