+++ rifGenericLibrary/src/main/java/org/sahsu/rif/generic/taxonomyservices/AbstractTaxonomyService.java
@@ -7,7 +7,9 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 import org.apache.commons.lang3.StringUtils;
 import org.sahsu.rif.generic.concepts.Parameter;
-import org.sahsu.rif.generic.fileformats.FilePath;
+import org.sahsu.rif.generic.fileformats.ClasspathFilePath;
+import org.sahsu.rif.generic.fileformats.tomcat.TomcatBase;
+import org.sahsu.rif.generic.fileformats.tomcat.TomcatFile;
 import org.sahsu.rif.generic.system.RIFServiceException;
 
 public abstract class AbstractTaxonomyService implements TaxonomyServiceAPI {
@@ -168,7 +170,7 @@ public abstract class AbstractTaxonomyService implements TaxonomyServiceAPI {
 			                              + "file", taxonomyServiceConfiguration.getName());
 		}
 
-		return new FilePath(taxonomyFileName).getPath();
+		return new TomcatFile(new TomcatBase(), taxonomyFileName).path();
 	}
 
 	/**
